Spread callbacks go to the object script
Previously, the spread callback would go straight to the FSM, bypassing
the object script. Now the spread signal first goes to the object,
passing the neighbour that called it, allowing for different behaviour
to be programmed for each object e.g. requiring different numbers of
neighbours to be on fire to catch itself.